errors.go moved to package book
